Moved functions from main.py into respective files. Args parse added. Preproc and descriptor directives are available
