+++ RESULTS.md
@@ -1,5 +1,6 @@
 # Results Gzip-based Molecular Classification
-|Data Set|Split |AUROC/RMSE (Valid)|F1/MAE (Valid)|AUROC/RMSE (Test)| F1/MAE (Test) |
-|--------|------|------------------|--------------|-----------------|---------------|
-|freesolv|random|0.456 +/- 0.094   |0.29 +/- 0.094|0.505 +/- 0.094  |0.308 +/- 0.094|
-|delaney |random|1.242 +/- 0.14    |0.967 +/- 0.14|1.359 +/- 0.14   |1.056 +/- 0.14 |
+|Data Set|Split |AUROC/RMSE (Valid)|F1/MAE (Valid) |AUROC/RMSE (Test)| F1/MAE (Test) |
+|--------|------|------------------|---------------|-----------------|---------------|
+|freesolv|random|0.588 +/- 0.244   |0.312 +/- 0.244|0.477 +/- 0.244  |0.305 +/- 0.244|
+|delaney |random|1.303 +/- 0.085   |1.005 +/- 0.085|1.349 +/- 0.085  |1.029 +/- 0.085|
+|lipo    |random|0.926 +/- 0.041   |0.738 +/- 0.041|0.926 +/- 0.041  |0.743 +/- 0.041|
