+++ classifiers.py
@@ -1,0 +1,5 @@
+from gzip_classifier import classify as gzip_classify
+
+classifiers = {
+    "gzip": gzip_classify,
+}
