+++ regressors.py
@@ -1,0 +1,6 @@
+from gzip_regressor import regress as gzip_regress
+
+regressors = {
+    "gzip": gzip_regress,
+}
+
